Return reference to last node from `find`
Instead of just returning whether the trie contains a given input, we
can provide a more flexible API by returning a reference to the `Node`
containing the last character of the input. This enables users to search
for a prefix, and continue searching from that prefix, instead of having
to search for the entire input each time.