Implement `Iterator` for `Trie`
This enables users to iterate over the words in a `Trie`, either from
the root of the `Trie`, or from a given prefix.

The `TrieRead` struct was added as a way to store the stack during
iteration. The `Trie` is iterated depth-first, and only complete words
are returned. When a `Node` doesn't have any children, it is the end of
a word, and is thus considered complete.

This is an initial implementation, and might not be the most optimal
solution.